Merge Onigmo 5.11.1.

Conflicts:
	.gitignore
	enc/ascii.c
	enc/euc_jp.c
	enc/shift_jis.c
	enc/utf_8.c
	oniguruma.h
	regcomp.c
	regenc.c
	regenc.h
	regerror.c
	regexec.c
	regint.h
	regparse.c
	regparse.h
	regsyntax.c
	st.h

+++ enc/utf_32le.c
@@ -159,7 +159,7 @@ utf32le_left_adjust_char_head(const UChar* start, const UChar* s, const UChar* e
 
   if (s <= start) return (UChar* )s;
 
-  rem = (s - start) % 4;
+  rem = (int )((s - start) % 4);
   return (UChar* )(s - rem);
 }
